+++ avs_server/src/lib.rs
@@ -15,7 +15,6 @@
 //
 
 pub mod ca;
-pub mod certs;
 pub mod config;
 pub mod csr;
 pub mod server;